backend/requirements.txt
# FastAPI Backend
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# Alpaca Trading API
alpaca-py==0.30.1

# Supabase Database
supabase==2.10.0

# OpenAI API for GPT-4o
openai==1.54.5

# ElevenLabs API for TTS
elevenlabs==1.10.0

# WebSocket support
websockets==13.1

# HTTP client
httpx==0.27.2

# Data handling
pydantic==2.10.3
python-multipart==0.0.18


